Only render MailList-component when needed
Previously we were passing `false` to the `MailList`-component under the
`mails`-prop and so the component would crash if any operations would be
done to the "array".
This commit fixes that and makes sure we only render the component when
we actually have an array of mails to show to the user.