Force submit and max length for year
Before this, a user could have a year with more than 4 digits.

We also showed a list of emails even if a user did not press the
`submit`-button for date-picker, now user has to submit.

Lastly, we log the selected dates to show that we can use it for sorting
emails out if we need to. To keep the different outputs of the date, we
won't do this though since we rely on dummy-data that will need a
changed date for every day this is tested.